+++ CHANGELOG.md
@@ -17,7 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
-*
+* Revert `node-fetch` bump back to @2
 
 ## [0.21.0] - 2022-02-15